64-bit transition part 13: the rest of ksh93 (#973)
This is the thirteenth of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space.

This patch applies the rest of the changes to ksh93 necessary for
64-bit memory allocation, along with the usual warning fixes.

Notable changes:
- parse.c: Fixed a build error when using SHOPT_KIA (re: 63366e34).
- sh_pipe(): Fix broken SOCK_CLOEXEC check. The !SOCK_CLOEXEC check
  in the #if directive apparently didn't work[*], as strace was
  showing unnecessary fcntl calls were still getting compiled in.
  Check for SOCK_CLOEXEC==0 in a non-preprocessor if statement and
  let the compiler optimize it away, which actually works
  correctly.
- Also removed some unnecessary ptrdiff_t casts.
- Add a (Linux-specific) regression test for allocating 5GB (we
  check if plenty of RAM is available via /proc/meminfo).
- locale.sh: Disable a test that requires SHOPT_MULTIBYTE to work
  correctly.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 352 => 37 (progression from part 12 =>
part 13)

The nvflags warnings that remain will be addressed in a separate
follow-up patch of their own (which doesn't affect memory
allocation).

Resolves: https://github.com/ksh93/ksh/issues/592

[*] (Martijn adds) On Linux, the SOCK_* macros defined in
/usr/include/bits/socket_type.h are bizarrely defined as their own
names (which refer to members of an enum type), which kills the #if
logic in the preprocessor. This could be an issue with other macros
on other systems too, so it's something to look out for in future.

+++ src/cmd/ksh93/sh/name.c
@@ -187,13 +187,13 @@ Namval_t *nv_addnode(Namval_t* np, int remove)
 			UNREACHABLE();
 		}
 	}
-	for(i=0; i < (unsigned)sp->numnodes; i++)
+	for(i=0; i < (size_t)sp->numnodes; i++)
 	{
 		if(np == sp->nodes[i])
 		{
 			if(remove)
 			{
-				while(++i < (unsigned)sp->numnodes)
+				while(++i < (size_t)sp->numnodes)
 					sp->nodes[i-1] = sp->nodes[i];
 				sp->numnodes--;
 			}
@@ -2894,7 +2894,7 @@ void nv_newattr(Namval_t *np, unsigned newatts, ssize_t size)
 				if(!*sp) sp--;		/* if number was 0, leave one zero */
 			}
 			len = strlen(sp);
-			cp = (char*)sh_malloc((len >= (unsigned)size ? len : (unsigned)size) + 1);
+			cp = (char*)sh_malloc((len >= (size_t)size ? len : (size_t)size) + 1);
 			strcpy(cp, sp);
 			if(sp && (mp=nv_opensub(np)))
 			{
